build-vm-image.sh: Add --no-libvirt option
Add a new command line option (`--no-libvirt`) that uses a direct call of QEMU
to run the Fedora installer instead of relying on virt-install and thus
not requiring the libvirtd to be installed.

Use QEMU and OVMF that we already built from scratch.

This allows running the script inside a podman container.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>