build-vm-image: Add c10s option
Add command line option to select the distribution to use for the guest
image; current options are f42 and c10s.

Do not silently overwrite/reuse the disk image if it already exists.
Add a new flag `--force` to overwrite it , if desired.

Change the target file name of the disk image to `guest-luks.qcow2`,
to be distro-neutral.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>